init pipeline
