+++ src/service/api.js
@@ -7,4 +7,53 @@ const api = axios.create({
   },
 });
 
+const refreshApi = axios.create({
+  baseURL: import.meta.env.VITE_API_URL,
+  headers: { 'Content-Type': 'application/json', Authorization: '' },
+});
+
+api.interceptors.request.use(
+  (config) => {
+    const accessToken = localStorage.getItem('accessToken');
+
+    if (accessToken) {
+      config.headers['Authorization'] = `Bearer ${accessToken}`;
+    }
+    return config;
+  },
+  (error) => {
+    return Promise.reject(error);
+  }
+);
+
+api.interceptors.response.use(
+  (response) => response,
+  async (error) => {
+    const originalRequest = error.config;
+
+    if (error.response.status === 401 && !originalRequest._retry) {
+      originalRequest._retry = true; // Mark the request to prevent infinite loops
+
+      try {
+        const refreshToken = localStorage.getItem('refreshToken'); // Retrieve refresh token
+
+        const response = await refreshApi.post('/auth/refresh-token', {
+          refreshToken: refreshToken,
+        });
+
+        const newAccessToken = response.data.accessToken;
+        localStorage.setItem('accessToken', newAccessToken); // Store new access token
+
+        originalRequest.headers['Authorization'] = `Bearer ${newAccessToken}`;
+
+        return api(originalRequest); // Retry the original request
+      } catch (refreshError) {
+        console.error('Refresh token failed:', refreshError);
+        return Promise.reject(refreshError);
+      }
+    }
+    return Promise.reject(error);
+  }
+);
+
 export default api;
